Move remaining circuit-core to circuit-foundation
