+++ circuit-codegen-annotations/src/commonMain/kotlin/com/slack/circuit/codegen/annotations/CircuitInject.kt
@@ -3,7 +3,7 @@
 package com.slack.circuit.codegen.annotations
 
 import androidx.compose.runtime.Composable
-import com.slack.circuit.CircuitConfig
+import com.slack.circuit.foundation.CircuitConfig
 import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.Screen
