+++ circuit-foundation/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=circuit-foundation
+POM_NAME=Circuit (Foundation)
+POM_DESCRIPTION=Circuit Foundation
