--- circuit/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=circuit-core
-POM_NAME=Circuit (core)
-POM_DESCRIPTION=Circuit
