--- circuit/src/commonMain/kotlin/com/slack/circuit/SaveableBackstack+Screen.kt
@@ -1,19 +1,0 @@
-// Copyright (C) 2022 Slack Technologies, LLC
-// SPDX-License-Identifier: Apache-2.0
-package com.slack.circuit
-
-import com.slack.circuit.backstack.SaveableBackStack
-import com.slack.circuit.runtime.Screen
-
-public fun SaveableBackStack.push(screen: Screen) {
-  push(
-    SaveableBackStack.Record(
-      route = screen.javaClass.simpleName,
-      args = mapOf("screen" to screen),
-      key = screen.hashCode().toString()
-    )
-  )
-}
-
-public val SaveableBackStack.Record.screen: Screen
-  get() = args.getValue("screen") as Screen
