+++ settings.gradle.kts
@@ -151,9 +151,9 @@ rootProject.name = "circuit-root"
 // Please keep these in alphabetical order!
 include(
   ":backstack",
-  ":circuit",
   ":circuit-codegen",
   ":circuit-codegen-annotations",
+  ":circuit-foundation",
   ":circuit-overlay",
   ":circuit-retained",
   ":circuit-runtime",
